test: extract visual float scope tests
